chore: remove stories cases
