+++ packages/datetime/src/components/DateTimeRangeFloatingPicker/DateTimeRangeFloatingPicker.stories.tsx
@@ -5,6 +5,8 @@ import { subMonths } from 'date-fns';
 
 import { DateTimeRangeFloatingPicker } from './DateTimeRangeFloatingPicker';
 import { onApply } from './__stories__/utils';
+import { useDefaultPresets } from '../Presets';
+import { TRealtimeState } from '../DateTimeRangeInput/declarations';
 
 const meta: Meta<typeof DateTimeRangeFloatingPicker> = {
   title: 'Components/Datetime/DateTimeRangeFloatingPicker',
@@ -15,6 +17,7 @@ const meta: Meta<typeof DateTimeRangeFloatingPicker> = {
       to: new Date().getTime(),
     },
     id: 'story-demo',
+    size: 'md',
   },
 };
 
@@ -37,4 +40,144 @@ export const Base: Story = {
         />
       );
     })(args),
+};
+
+export const DefaultPresets: Story = {
+  render: (args) =>
+    ((props) => {
+      const [date, setDate] = React.useState(props.value);
+
+      const presets = useDefaultPresets();
+
+      return (
+        <DateTimeRangeFloatingPicker
+          {...props}
+          value={date}
+          onApply={onApply(setDate)}
+          onChange={(range) => {
+            console.log('something has changed', range);
+          }}
+          onRealTimeClick={(event) => {
+            console.log('RT button clicked', event);
+          }}
+          presets={presets}
+          parseExpression={(exp) => {
+            for (const word of ['now', 'today', 'yesterday']) {
+              if (typeof exp === 'string' && exp.includes(word))
+                return {
+                  isValid: true,
+                  value: new Date().getTime(),
+                  errors: [],
+                };
+            }
+            return {
+              isValid: false,
+              value: null,
+              errors: ['expression no valid'],
+            };
+          }}
+        />
+      );
+    })(args),
+};
+
+export const CustomPresets: Story = {
+  render: (args) =>
+    ((props) => {
+      const [date, setDate] = React.useState(props.value);
+      const [realtime, setRealtime] =
+        React.useState<TRealtimeState>('disabled');
+      const [btnDisabledApply, setBtnDisabledApply] = React.useState(
+        props.disableApplyButton,
+      );
+
+      const realtimeCallback = React.useCallback(() => {
+        if (realtime === 'inactive') {
+          setRealtime('activated');
+          setDate({ from: '', to: '' });
+        } else if (realtime === 'activated') {
+          setRealtime('inactive');
+          setDate(props.value);
+        }
+      }, [props.value, realtime]);
+
+      const onChangeCallback = React.useCallback(
+        (range: {
+          timestamp: { from: number; to: number };
+          preset: { from: string; to: string };
+        }) => {
+          // control Apply button
+          if (
+            (range.timestamp.from && !range.timestamp.to) ||
+            (range.timestamp.to && !range.timestamp.from) ||
+            (range.preset.from && !range.preset.to) ||
+            (range.preset.to && !range.preset.from)
+          ) {
+            setBtnDisabledApply(true);
+          } else {
+            setBtnDisabledApply(false);
+          }
+
+          // control realtime button
+          if (range.timestamp.from || range.timestamp.to) {
+            setRealtime('disabled');
+          } else {
+            setRealtime('inactive');
+          }
+        },
+        [],
+      );
+
+      return (
+        <DateTimeRangeFloatingPicker
+          {...props}
+          value={date}
+          onApply={onApply(setDate)}
+          onChange={onChangeCallback}
+          onRealTimeClick={realtimeCallback}
+          presets={[
+            {
+              label: 'Relative too',
+              options: [
+                {
+                  value: {
+                    from: 'now() - 30s',
+                    to: 'now()',
+                  },
+                  label: 'Last 30 seconds',
+                },
+                {
+                  value: {
+                    from: 'now() - 1m',
+                    to: 'now()',
+                  },
+                  label: 'Last minute',
+                },
+              ],
+            },
+          ]}
+          parseExpression={(exp) => {
+            for (const word of ['now', 'today', 'yesterday']) {
+              if (typeof exp === 'string' && exp.includes(word))
+                return {
+                  isValid: true,
+                  value: new Date().getTime(),
+                  errors: [],
+                };
+            }
+            return {
+              isValid: false,
+              value: null,
+              errors: ['expression no valid'],
+            };
+          }}
+          realTime={realtime}
+          disableApplyButton={btnDisabledApply}
+        />
+      );
+    })(args),
+  args: {
+    realTime: 'inactive',
+    showCalendarIcon: true,
+  },
 };
